+++ src/App.vue
@@ -1,26 +1,9 @@
 <script setup lang="ts">
-import { onMounted } from 'vue';
-
-onMounted(() => {
-  document.body.classList.add('fade-in');
-});
+import MainLayout from './components/layout/MainLayout.vue';
 </script>
 
 <template>
-  <router-view />
+  <MainLayout>
+    <router-view />
+  </MainLayout>
 </template>
-
-<style>
-.fade-in {
-  animation: fadeIn 1s ease-in;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-</style>
